Move RS256AlgorithmTests into an 'Algorithms' sub-folder
